first round of tests (for .js/.cjs/.mjs)
